+++ pkgs/timezone/lib/standalone.dart
@@ -16,23 +16,22 @@ library timezone.standalone;
 
 import 'dart:async';
 import 'dart:io';
-import 'dart:mirrors';
+import 'dart:isolate';
 import 'package:path/path.dart' as path;
 import 'package:timezone/timezone.dart';
 
 export 'package:timezone/timezone.dart'
     show getLocation, setLocalLocation, TZDateTime, Location, TimeZone, timeZoneDatabase;
 
-final _packagesPrefix = 'packages${path.separator}';
-
-final String tzDataDefaultPath = path.join('packages', 'timezone', 'data', tzDataDefaultFilename);
+final String tzDataDefaultPath = path.join('data', tzDataDefaultFilename);
 
 // Load file
-Future<List<int>> _loadAsBytes(String p) {
+Future<List<int>> _loadAsBytes(String p) async {
   final script = Platform.script;
   final scheme = Platform.script.scheme;
 
   if (scheme.startsWith('http')) {
+    // TODO: This path is not tested. How would one get to this situation?
     return new HttpClient()
         .getUrl(new Uri(scheme: script.scheme, host: script.host, port: script.port, path: p))
         .then((req) {
@@ -43,37 +42,10 @@ Future<List<int>> _loadAsBytes(String p) {
         return builder.takeBytes();
       });
     });
-  } else if (scheme == 'file') {
-    final packageRoot = Platform.packageRoot;
-    if (packageRoot != null && packageRoot.isNotEmpty && p.startsWith(_packagesPrefix)) {
-      p = path.join(path.fromUri(packageRoot), p.substring(_packagesPrefix.length));
-      return new File(p).readAsBytes();
-    }
-
-    p = path.join(path.dirname(path.fromUri(script)), p);
-    return new File(p).readAsBytes();
-  } else if (scheme == 'data') {
-    var libraryPath = currentMirrorSystem().findLibrary(#timezone.standalone).uri.path;
-    var prefix = path.join(_packagesPrefix, path.dirname(libraryPath));
-    p = path.join(prefix, p.substring(prefix.length + 1));
-    return new File(p).readAsBytes();
+  } else {
+    var uri = await Isolate.resolvePackageUri(new Uri(scheme: 'package', path: 'timezone/$p'));
+    return new File(path.fromUri(uri)).readAsBytes();
   }
-
-  return new Future.error(new UnimplementedError('Unknown script scheme: $scheme'));
-}
-
-List<int> _loadAsBytesSync(String p) {
-  assert(!Platform.script.scheme.startsWith('http'));
-
-  final script = Platform.script;
-  final packageRoot = Platform.packageRoot;
-  if (packageRoot != null && packageRoot.isNotEmpty && p.startsWith(_packagesPrefix)) {
-    p = path.join(path.fromUri(packageRoot), p.substring(_packagesPrefix.length));
-    return new File(p).readAsBytesSync();
-  }
-
-  p = path.join(path.dirname(path.fromUri(script)), p);
-  return new File(p).readAsBytesSync();
 }
 
 /// Initialize Time Zone database.
@@ -97,27 +69,4 @@ Future initializeTimeZone([String p]) {
   }).catchError((e) {
     throw new TimeZoneInitException(e.toString());
   });
-}
-
-/// Initialize Time Zone database (Sync).
-///
-/// Throws [TimeZoneInitException] when something is worng.
-///
-/// ```dart
-/// import 'package:timezone/standalone.dart';
-///
-/// initializeTimeZoneSync();
-/// final detroit = getLocation('America/Detroit');
-/// final detroitNow = new TZDateTime.now(detroit);
-/// ```
-void initializeTimeZoneSync([String p]) {
-  if (p == null) {
-    p = tzDataDefaultPath;
-  }
-  try {
-    final rawData = _loadAsBytesSync(p);
-    initializeDatabase(rawData);
-  } catch (e) {
-    throw new TimeZoneInitException(e.toString());
-  }
 }